+++ app/db/service/resource_types/resource_type.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+require 'singleton'
+
+module DB
+  module Service
+    module Types
+      class ResourceType
+        include Singleton
+
+        def create(resource)
+          raise NotImplementedError
+        end
+
+        def delete(resource)
+          raise NotImplementedError
+        end
+      end
+    end
+  end
+end